% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arudata.R
\name{aru_repo}
\alias{aru_repo}
\title{Listado de bases de datos disponibles}
\usage{
aru_repo()
}
\description{
Esta función desplega las bases de datos disponibles
}
\details{
Versión piloto
}
\examples{
aru_repo()
}
