% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arudata.R
\name{arudata}
\alias{arudata}
\title{Cargar base de datos}
\usage{
arudata(id = "cs76v", pass = 9876)
}
\arguments{
\item{id}{Es el identificador de la base de datos}

\item{pass}{Es la contraseña para acceder a la base de datos definida en id}
}
\description{
Esta función permite cargar las bases de datos alojadas en los servidores de Fundación Aru.
}
\details{
Versión piloto
}
\examples{
arudata(id="cs76v",pass=9876)
}
