% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{indexaru}
\alias{indexaru}
\title{13 PIACC variables for SPAIN}
\format{An object of class \code{data.frame} with 4 rows and 3 columns.}
\usage{
indexaru
}
\description{
PIACC data for SPAIN. PIACC es el acrónimo de
}
\keyword{datasets}
